clean the socket when close connect
